feat(exchange-rate): 환율 조회 방식을 SSE에서 Polling으로 전환 (커넥션 최적화)-#50
[Architecture Decision]
- 주식 거래 및 실시간 호가창 전송에 SSE 커넥션이 집중될 예정이므로, 환율까지 SSE를 유지할 경우 사용자당 다중 커넥션이 발생함.
- 이로 인한 서버의 커넥션 풀 고갈 및 TCP Sequence Number 고갈 위험을 사전에 방지하기 위해 환율 도메인을 REST API 기반의 Polling 방식으로 전환.
