+++ tradelink/src/main/java/site/tradelink/tradelink/exchangeRate/repository/CurrentExchangeRateCustomRepository.java
@@ -1,0 +1,9 @@
+package site.tradelink.tradelink.exchangeRate.repository;
+
+import site.tradelink.tradelink.exchangeRate.entity.CurrentExchangeRate;
+
+import java.util.List;
+
+public interface CurrentExchangeRateCustomRepository {
+    List<CurrentExchangeRate> findAll();
+}
