+++ tradelink/src/main/java/site/tradelink/tradelink/exchangeRate/repository/DailyExchangeRateRepository.java
@@ -1,0 +1,10 @@
+package site.tradelink.tradelink.exchangeRate.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import site.tradelink.tradelink.exchangeRate.entity.DailyExchangeRate;
+
+@Repository
+public interface DailyExchangeRateRepository extends JpaRepository<DailyExchangeRate, Long>, DailyExchangeRateCustomRepository {
+
+}
