+++ tradelink/src/main/java/site/tradelink/tradelink/exchangeRate/repository/ExchangeRateRepository.java
@@ -1,0 +1,10 @@
+package site.tradelink.tradelink.exchangeRate.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import site.tradelink.tradelink.exchangeRate.entity.ExchangeRate;
+
+@Repository
+public interface ExchangeRateRepository extends JpaRepository<ExchangeRate, Long>, ExchangeRateCustomRepository {
+
+}
