--- tradelink/src/main/java/site/tradelink/tradelink/stock/repository/CurrentExchangeRateCustomRepository.java
@@ -1,9 +1,0 @@
-package site.tradelink.tradelink.stock.repository;
-
-import site.tradelink.tradelink.stock.entity.CurrentExchangeRate;
-
-import java.util.List;
-
-public interface CurrentExchangeRateCustomRepository {
-    List<CurrentExchangeRate> findAll();
-}
