--- tradelink/src/main/java/site/tradelink/tradelink/stock/repository/DailyExchangeRateRepository.java
@@ -1,13 +1,0 @@
-package site.tradelink.tradelink.stock.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import site.tradelink.tradelink.stock.entity.DailyExchangeRate;
-
-import java.time.LocalDate;
-import java.util.List;
-
-@Repository
-public interface DailyExchangeRateRepository extends JpaRepository<DailyExchangeRate, Long>, DailyExchangeRateCustomRepository {
-
-}
