--- tradelink/src/main/java/site/tradelink/tradelink/stock/sse/dto/SseEvent.java
@@ -1,12 +1,0 @@
-package site.tradelink.tradelink.stock.sse.dto;
-
-public record SseEvent(String eventName, Object data) {
-
-    public static SseEvent poison() {
-        return new SseEvent("__CLOSE__", null);
-    }
-
-    public boolean isPoison() {
-        return "__CLOSE__".equals(eventName);
-    }
-}
